Add endpoints for orders and products

+++ src/products/products.controller.ts
@@ -1,8 +1,17 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  NotFoundException,
+  Param,
+  Post,
+  Put,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
-import { Product } from 'src/db';
 import { CreateProductDTO } from './dtos/create-product.dto';
 import { ParseUUIDPipe } from '@nestjs/common';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -14,15 +23,34 @@ export class ProductsController {
   }
   @Get('/:id')
   getById(@Param('id', new ParseUUIDPipe()) id: 'string'): any {
-    return this.productsService.getById(id);
+    const prod = this.productsService.getById(id);
+    if (!prod) {
+      throw new NotFoundException('Product not found');
+    } else {
+      return prod;
+    }
   }
   @Delete('/:id')
   deleteById(@Param('id', new ParseUUIDPipe()) id: 'string'): any {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
     this.productsService.deleteById(id);
     return { success: true };
   }
   @Post('/')
   create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
+  }
+  @Put('/:id')
+  edit(
+    @Body() newProductData: UpdateProductDTO,
+    @Param('id', new ParseUUIDPipe()) id: 'string',
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
+    this.productsService.editProduct(newProductData, id);
+    return { success: true };
   }
 }
